requirements.txt
Plaintext
fastapi
uvicorn
psycopg2-binary
google-generativeai
python-dotenv